requirements.txt
graphics.py==5.0.0
PyOpenGL==3.1.6
numpy==1.21.0